+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 ## 0.0.2
 * Update of the comments inside our package.
 * Updates inside pubspec.yaml
+
+## 0.0.3
+* Update of the comments inside our package.
+* Update to the main.dart file inside the example folder.